requirements.txt
streamlit
pypdf
langchain
python-dotenv
chromadb
langchain-google-genai==2.1.0